feat(og): load real Geist faces, split the peg card, tighten its stale window
Closes the three deferrals opened against the peg-monitoring social card.

**Fonts (#1920).** All four cards declared `fontFamily: "Geist"` and rendered
in Satori's bundled fallback, because none passed a `fonts` array. The repo
only ships `geist-latin.woff2` and Satori cannot parse woff2 at all, so this
adds the plain-woff cuts of the same family (400/600/700/800, ~40KB each) and
a shared `ogFontOptions()` loader.

Two details the render caught that reading would not have. `fetch` cannot read
a `file:` URL on the Node runtime these routes use, so the faces load through
`readFile` over a `new URL(..., import.meta.url)` — still the form the bundler
traces, and the four woff files now appear in every card's `.nft.json`, so
Vercel ships them into the function. And the loader returns `{}` rather than
`{ fonts: [] }` on failure: an empty array is a font set with nothing in it,
which Satori rejects outright, turning a cosmetic fallback into a 500.

With weights no longer inert, the peg card gets its hierarchy back — pair,
price, status and the verdict pill carry 600/700 instead of leaning on size
and colour alone. The other three cards' existing weight declarations start
working on their own.

**Split (#1921).** The card was ~670 lines against a 600-line soft cap. Its
Satori JSX moves to `_og/{theme,rail,tile,table}`, leaving the route file at
223 lines holding only the card frame and entry points. `_lib` is now the
non-JSX layer and `_og` the render layer, which also makes it obvious that
nothing from `_components/` (all `"use client"`) may enter this import graph.

**Stale window (#1922).** The peg card drops to
`stale-while-revalidate=300` where the other three keep 86400. Those carry TVL
and volume, where stale means out of date; this one carries an alert
condition, and a 24h window lets a CDN answer a request made hours into a
breach with a card rendered while every peg was healthy. The cost is a real
render (~1-2s) for an unfurl after a quiet spell.

Renders verified against the real route for the single-peg tile and the
four-peg table, before and after the split, plus a production build confirming
the font traces into the serverless output.

Closes #1920
Closes #1921
Closes #1922
