+++ ui-dashboard/src/app/opengraph-image.tsx
@@ -1,4 +1,5 @@
 import { ImageResponse } from "next/og";
+import { ogFontOptions } from "@/lib/og-fonts";
 import { fetchHomepageOgData, type HomepageOgData } from "@/lib/homepage-og";
 import { formatUSD } from "@/lib/format";
 
@@ -276,6 +277,7 @@ export default async function Image() {
   const data = await fetchHomepageOgData();
   return new ImageResponse(<Card data={data} />, {
     ...size,
+    ...(await ogFontOptions()),
     headers: { "Cache-Control": IMAGE_CACHE_CONTROL },
   });
 }
